Fix bug: inquire-read average phi only if restart
If the moethod chosen for initial bed roughness is not "restart",
no need to inquire/read the averaged inverted bed roughness from
the restart file. Otherwise, this could cause problems when a
restart is wanted from a previous run that did not invert for
bed roughness; in that case ave phi was never written to output.