Add the option to force a mesh creation
Now the model can run a mesh update after a specified
number of years since the beginning of a run, even if
the mesh fitness is good enough. This allows reducing
the mesh resolution if, for any reason, we want to do
it. Useful for restarting from a nut-high calibration
run, for example.