Add exception for margin vertices: no upwind triangle found
Margins are complicated. Sometimes the thermodynamics module
can't find where the ice is coming from for a marginal vertex.
The default option in this case is to crash. Made an exception
so in this marginal cases the multiplication between the ice
temperature gradients and ice velocity is 0, instead of a painful
crash.