Add exception for margin triangles: no b-grid velocity found
At the border, sometimes the mapping of the 3D velocities between
the b- and c-grids fails to find the triangle from which the vels
are taken, and the model crashes. Added an exception to simply set
the velocity there to 0, as these points are not that important
(or at least not as important as preventing a total model crash).