Add slight memory method difference restart vs new run
New runs use a much milder memory method, that linearly increases
the weights for senility in order to speed up the first few years
of the simulation. Restarts, on the other hand, use a much stronger
memory, which allows new rates of ice change to kick in at the end
of the memory/senility period of activity.

Added a few safety nets to the computation of the critical SIA and
adv time steps, to make sure that they are not smaller than the
minimum time step allowed. This helps with the "direct" time-stepping
method. However, it is still not running perfectly. WIP.