Fix remapping of bed roughness after mesh update
Classic information problem. After a mesh update, the bed
roughness is mapped to the new mesh, but since there was no
"reference" roughness field, information was lost more and more
after each mesh update. Now, the bed roughness is mapped from the
restart data directly each time, which is assumed to be at a high
enough resolution to cover new areas of high mesh density. Further,
this mapping is done in the LOG domain to compensate for differences
in the orders of magnitude between the friction and sliding values.