Bring back precipitation interpolation on LOG domain
For tests, the interpolation of precipitation snapshots for the
climate matrix was being done linearly. This does not account
for the differences in orders of magnitude in the precipitation
rates between the snapshots. Now done in the LOG domain as before,
to account for this, while making sure that precipitation for full
warm periods is equal to PD up to computer precision.