Improve mapping of bed roughness after mesh update
Now the mapping is done using two different methods, which
are then combined to compensate for interpolation errors.
The result is further fine-tuned by adjusting the bed roughness
within the range generated by the two mapping methods, trying
to minimise the difference between the previous dHi_dt history
and the new rates of change after the remap.