Add use of average bed roughness beyond inversion period
Once the inversion reaches the end of the adjustment period,
the model will continue the simulation using the averaged
inverted values over the last running window. This prevents
bumps in the ice volume trend caused by simply using the
last "snapshot" of the inversion, which contains an artificial,
"instantaneous" trend.