Add start-up / cool-down phases for better restarts
At the start and end of any simulation, the time step is
smoothly reduced to the minimum value allowed, and the
tolerance for the velocity solution is also reduced. Since
this is also done for a restart run, there is a better
agreement between the previous and restart run velocity-
solver setups, which reduces the initial restart shock.