Make time-step depend only on ice dynamics
Back to the basics. Now the other submodels and output
wait until the region time is greater than their next
update time to fire up. Not as elegant as before but
at least the time step is not affected by choices of
other time steps anymore. Still, the time step depends
on the coupling interval and end of run.