Add update of prescribed sea level every time step
This way the use of a prescribed sea level record can be
used independently of the coupling interval, which is really
needed only when coupling the ice sheets through the eustatic
sea level method. This way, a single-domain simulation can be
run with no coupling interval at all, for example.