Update memory/senility method to work with Hi
Instead of dHi_dt. Furthermore, now the weighed average
is computed between the previous and current time step,
for a smoother transition. As the time step reduction works
alright for restarts, this memory method is better suited
for controlled calibrations, to avoid a super big bump at
the start when the bed roughness coefficients are just 0s.