Add option to use single process for mesh generation
As of now, mesh generation on multiple processes suffers
from a limitation on the max resolution as the number of
processes increases. In order to seize the bigger number
of cores available on supercomputers, added an option to
use a single process for mesh generation, and the full
number of processes for the rest of the model.

Another missing call to sync in the wind-up phase added.