Add use of ice masks from GCM output data
Now the model uses the mask info from the GCM directly (which
should be standard output for GCMs anyway), instead of using
different proxies to guess the ice distribution in the model
to get some of the climate state variables, e.g. total absorbed
insolation references. Rewrote the precipitation interpolation
part of the climate matrix to use these masks instead.