Remove very thin ice in BedMachine data
Added the option to remove very thin ice from reference geometry. Particularly for BedMachine Greenland this can be an issue as for reasons unknown it covers almost the entire tundra with half a metre of ice. This messes up the definition of the present-day ice mask, which is now fixed.