Add main model loop
Added all the basic stuff needed to run a simulation
from start to end, i.e. most of the routines in the
run_model subroutine. The model compiles and runs,
but crashes often due to instabilities coming from
ill-solved velocities (for any number of processes).
Thus, next step is a detailed revision of the code
flow from the start until the computation of velocities,
including a comparison to the develop branch.