Add initial version of grounded area fractions
Computed from the high resolution bedrock reference data,
and used to scale the friction term beta_eff that enters
the velocity solver, plus in the inversion of basal melt
to identify places where calving and melt occur at sub-grid
scales. Needs to be upgraded to the density functions and
to the use of UFEMISM's native functions to compute the
number of sub-grid cells that fall within each triangle.