Add modified bed inversion based on hydro proxies
Now the inversion of bed roughness is modulated by the
expected presence of liquid water at the base, which in
the model now depends on 1) basal ice temps, 2) bedrock
elevation above sea level, and 3) percolated runoff.
For example, for areas where the ice base is very cold,
the runoff is zero, and the bedrock is well above sea
level, the inversion will be reduced to almost nothing.
This helps preventing inversion over e.g. mountaineous
regions, where the model kinda breaks if velocities
are too high.