Fix bug: missing grid values for high-res ocean
When extrapolating new ocean data into the ice shelf
cavities, some of the paramaters for the high-resolution
regular grid were allocated and initialised twice. Further,
when reading the extrapolated data from a file in subsequent
simulations, some of the parameters were missing, which
made the mapping from the grid onto the mesh kaput. Fix all
this.