Wrote flexible input file code
New code is added but not yet used for flexible reading of input files.

The subroutine "read_field_from_file_2D" in the new (unused) NetCDF module can be called, with a file name (of a NetCDF file) and a list of possible data field names as input. The file can contain data on an x.y-grid, a lon/lat-grid, or a mesh - the code will decide this for itself based on the dimensions that are present in the file. It will then look for a variable matching any of the provided field name options, throwing an error when no matches, or more than one match, are found. It will map the data from whatever grid it is provided on by the file, to the model mesh, so that the output from "read_field_from_file_2D" is on the model mesh.

When combined with the upcoming Atlas Update, this will make handling input data a walk in the park.