Added time dimension
Added support for reading data as a timeframe.

The main subroutine "read_field_from_file_2D" can now be provided an optional argument "time_to_read". If that argument is present, it assumes the data field in the file has a time dimension, and the timeframe closest to time_to_read should be read. If that closest match is not an exact match, a warning will be displayed. If the data in the file has a time dimension when none is expected, or vice versa, the model will throw an error.

Also restructured the routines a little so all error messages will display the full name of the file where the error was encountered.