Fix bug: skip interpolation for start/end of forcing records
When the simulation time is exactly the start or end of one of
the forcing records (e.g. t=0.0 and CO2), the forcing subroutines
were still interpolating, when they could have simply taken the
number directly from one of the extremes. This caused a few
ugly issues on Snellius. Fixed this by taken the extreme value
directly from the record.