Update masks determination + update ice model remap
A few changes to the determination of model masks:
- New grounding line mask at the floating side, i.e.
the first ice shelf point next to a "grounded-side"
grounding line.
- Moved the communication of basic masks (land, ocean,
ice, sheet, shelf) to before the determination of the
transitional masks (coast, margin, GL, calving front),
so the former can be used during the computation the
latter.

Updated the remap_ice_model routine a bit to remove
a couple of double bookings. This will be revisited
again anyway when I reach the mesh update part.