CHECKPOINT: Ocean + restart + equilibrium works
After a calibration, including the inversion of bed roughness
and ocean temperatures that operate until a steady-state is
reached, a restart is now possible using the inverted values
as constant fields, which manages to keep the restarted ice sheet
in equilibrium. The "PMP" BMB method keeps the steady-state
particularly well, after using the new sub-grid grounded
fractions. The ocean and reference_fields modules still
contain several lines of WIP code, which need to be cleaned up,
thus the checkpoint.