Tweak ocean inversion: no cf, bit of freeze, scale
Left the calving front out of the inversion, as it
would account for calving as well and that's something
we gotta deal with in the calving module. Instead,
inverted values are extrapolated from the interior.

Increased the inversion scaling, so adjustments are
smaller each inversion step. This should eventually
be a flexible config option.

Allowed for a slightly "too cold" ocean temperatures,
below the freezing point, to get some refreezing where
necessary.