Add fixed/delayed evolution of ice thickness
Updated the options for fixed geometry, so now instead of
a logical flag yo can input a real number between 0 and 1.
If the number for a given region is 0, it means that it is
not fixed at all. If it is 1, it is fully fixed. Any number
in-between will delay the evolution of ice thickness by
computing a weighed average between the new ice thickness
and the old one from the previous time step. The latter is
quite usefule for the initialisation of very unstable
geometries, such as that in Antarctica.