WIP: Revamp bed rough and ocean temp inversions (stable ANT: gsym)
Added a subgrid grounded area fraction based on Tijn's cumulative
density function approach. Kept the old one for reasons.

Standarised the names of config options for both inversions:
slid_inv and ocean_inv. Did the same for the clim_inv one btw.

Changed the way the scaling factor interacts with the inversion
adjustment for both ocean_inv and slid_inv: now the initial
inverse of the thickness scale (optionally) value decreases over
time, ever reducing the magnitude of the adjustment between the
start (full adjustment) and end (min adjustment) of the inversion
period. This reduces the oscillating overshoots caused by too-big
time steps or adjustments, leading to a smooth transition between
inverision and free-run. WIP.

Added a new tech where at the final section of an inversion, only
positive thickness misfits are dealt with, this further prevents
oscillations and helps the ice sheet reach a steady-state.

-----
Got a stable ANT domain under PD conditions. Saving this checkpoint.
-----

Added the option to keep the mask fixed, kind of: now the ice
thickness for floating ice can be limited to never exceed the
floatation thickness, effectively keeping it floating. Likewise
for the grounded ice: never let it get thinner than the floatation
thickness. This way the thickness ice never fixed, but the inversions
can work on the domain without the the grounding lines migrating like
crazy. This works, but once the fixing of the mask is removed, things
get out of control again. Seems like the inversions relay too much on
this tech, and their inverted parameters are not solid enough to
account for the missing constrain once it is removed. WIP.