WIP: Revamp bed rough and ocean temp inversions (stable ANT: cra1)
Introduced assymetric inversion: too think ice receives a stronger
adjustment than too thin ice for bot inversions. This allows to still
compensate for too thin ice near the end of the inversion, reducing
ovrshooting a bit. For sliding, too thin ice is adjusted for at a
1-to-3 ratio compared to too thick ice, during the entire inversion.
For ocean, the ratio is 1-to-10, but applied only during the last
stage of stabilisation. New config file that generates a stable
ANT is added (_cra1).