Velocity overhaul
A big, big update to the ice velocity, ice dynamics, and thermodynamics module(s).

Rewrote the velocity module to be completely modular with respect to different stress balance approximations; until now, a lot of hard-coded stuff was specific to the hybrid SIA/SSA. That is now fixed.

Rewrote the entire SIA, SSA, and DIVA solvers to be much more clean, understandable, and efficient.

Added a new module for the Blatter-Pattyn Approximation. Verified that it works in the ISMIP-HOM experiments, and in EISMINT1_A. Preliminary tests with EISMINT1_A show that running the model with the BPA instead of DIVA makes it roughly 10 times slower.

Rewrote most of the mesh_operators_module, where the matrices representing gradient/mapping operators are calculated. They now contain nice PETSc-based matrix operations to go from the [x',y',zeta]-coordinate operators to the geometry-dependent [x,y,z]-coordinate operators. The routines for calculating the shape functions are also cleaned up, and do now (in principle) allow for higher-order shape functions, although these are not used anywhere (yet).

Fixed a small bug in the DIVA, where the basal friction coefficient beta_b was calculated using u_vav rather than u_base. This did not show up in the ISMIP-HOM experiments because those have a velocity-independent friction coefficient.

Rewrote the thermodynamics solver to a more cleanly matrix-based code. Greatly extended the comments to make it more understandable. Also added a semi-dynamic time step. The solver will first "try" the config time step. If that gives a seemingly unstable solution (for now determined by the same conditions as before, i.e. T < 180 K or T = NaN), the time step is reduced by half, and the solver tries again. This is repeated until a stable solution is found, or until the time step becomes very small (right now 10 iterations, so 1/1024th of the config time step). In that case, the Robin solution is used instead.

Rewrote the vertical velocity calculation, as the old version was specific to an SIA-only model, and therefore incorrect. The new version explicitly solves for conservation of mass, using the same finite volume formulation as the ice thickness update. This fixes the problem of unphysical upward ice motion near the margins, and seems to also improve the stability of the thermodynamics solver.

Rewrote the strain heating calculation, as the old version was specific to an SIA-only model, and therefore incorrect. The new version now follows the general strain-rate-based formulation by Bueler and Brown (2009). The strain rates are provided by the different stress balance solvers. Strain rates that are neglected in the derivation of a certain stress balance approximation (e.g., the SIA neglects du/dx, du/dy, dv/dx, dv/dy, while the SSA neglects du/dz, dv/dz), are set to zero by the solvers.

Fixed a small bug in the mesh-to-grid remapping; grid cells that lie entirely outside of the mesh domain are now assigned a value by nearest-neighbour extrapolation.

Slightly changed the new NetCDF code. Having every single safety check separately open the file, check it, and then close it again made everything super slow. Instead, opening and closing the file is now once again done by the main read/write routines; all intermediate and lower routines are simply handed the ncid. With this change, writing to restart/help_fields/debug files once again doesn’t take any noticeable time.

Known issues:
1) cannot write to output files (restart/help_fields) when they are open in ncview.
2) DIVA still pancakes.