Bugfixes
Fixed some minor issues:

Remapping: added an exception in mesh-to-mesh remapping for the trivial case where the meshes are identical (which is the case when restarting the model with a previous simulation).

Fixed the vertical velocity calculation again. It now once again gives results matching analytical solutions in the EISMINT1 experiments, and now includes in the comments a step-by-step derivation of the equations.

Removed the code that set basal ice temperatures for floating ice equal to the ocean temperature at that depth; since the ice will always already be at pressure melting point due to the sliding at the grounding line, and since the ocean cannot be colder than pressure melting point, just set the temperature equal to PMP.

Fixed a small parallelisation bug in the code that determines the output file names, which caused a racing condition that sometimes crashed the model.

Rewrote the subroutines for calculating SMB and temperature in the EISMINT1 experiments to be closer to the equations in Huybrechts et al. (1996).