Add lateral melt for grounded calving fronts
Added an additional term in the mass balance that uses the
BMB from the ocean around a grounded calving front to compute
a lateral melt. This melt is proportional to the face of the
ice front that is underwater.

Also the effective thickness ice front fraction is now computed
only from surrounding ice shelf points (for floating calving fronts)
or only from surrounding ice sheet points (for grounded fronts).
Experimental, and prolly does not work with the threshold calving
parameterisation and new ice shelf births, but we gotta get rid of
that parameterisation eventually anyway. WIP