Fix bug: too coarse bed inversion extrapolation radius
The routine call used for the extrapolation of the inverted bed
roughness had by default a sigma of 10000, which was to small
for simulations using max resolutions of 80 km or coarser.
This created NaNs in the inverted field, which then made the
velocities kaput. Fixed this by increasing the radius to 40000.

Also, added the safety check for NaNs in the phi_fric and beta_sq
fields at the end of the inversion.