Fixed surface boundary conditions in the BPA solver
Fixed the implementation of the Neumann boundary conditions at the ice surface in the Blatter-Pattyn Approximation.

By using two-sided differencing with a ghost point above the surface, and substituting the resulting expressions for du/dz, d2u/dz2 into the BPA itself, the solution becomes a lot more accurate.

In the SIA solver, this finally leads to second-order convergence, and very little dependence on the vertical grid spacing ratio R. For the BPA I cannot test the convergence, but when using nz = 12, all ratios of 0.1 <= R <= 10 give results within the full-Stokes range for the ISMIP-HOM_A_160 experiment. nz = 8 still gives good results when 0.2 <= R <= 5; nz = 8 and R = 10 overestimates the fast part of the velocity by about 20 %.