Add rough sea init for sliding inversion
Now the ocean starts with the uniform values (as opposed
to the minimum angle value). This allows for an easier ice
sheet explansion during a glacial cycle. Sliding values can
then be fine-tuned during deglaciation where needed.

This requires an extrapolation of the grounded friction values
over every floating side of the grounding line, and not only
those that are partially grounded, since the assumption of full
sliding there is no more.