Update asymmetry in sld and ocn inversions
Removed the asymmetry in the sld inversion, i.e. now both thickening
and thinning are dealt with in the same manner throughout the inversion.
Updated the asymmetry in the ocn inv: Now the adjustment scales are kept
different for thick- and thinning, but the adjustment itself is the same
for both cases.

Removed the special treatment of the Amundsen area regarding beta/_eff.