Tweak sld and ocn inversions: no Siple and asymmetry
The inversion of bed roughness now is now asymmetric in
the thickness vs thinning decrease in the adjustment over time:
while adjustments for positive misfits transition from the initial
value for the inversion scale to the final one, the adjustments for
negative misfits go from initial scale to 0. This reduces the GL
advance once an equilibrium simulation is freed from inversions.

Added an exception for the Siple Coast area where it meets the
Transantarctic Mountains, which is very very sensitive to even
small overestimations of melt. Now no melt is allowed in that
small area during the inversion of ocean temperatures.